CA-206396: InstallAgent raises exception on opening reg key
This behaviour is only seen on the driver upgrade case. After the first
reboot, InstallAgent is able to read the state of the installer from
the registry and continues where it left off (uninstalling the MSIs).
However, uninstalling 'Citrix XenTools Installer' takes ~10 minutes
to complete. After 'UninstallMSIs()' returns, we try to set the
'MSIsUninstalled' flag to the registry and an exception is raised
because it cannot be opened.

This commit comprises 2 possible fixes:
- Wait until the '_MSIExecute' mutex is available/free before
  rebooting (if AUTOREBOOT)
- Enclose the install state reg key in a 'using' statement in the
  'Installer' static constructor

Signed-off-by: Kostas Ladopoulos <konstantinos.ladopoulos@citrix.com>