CA-213907 : ensure lite agent WMI connection is reestablished
Following a reinstall of the same version of xeniface, the lite agent
fails to notice its WMI agent is broken.  As a workaround, having
installed Xeniface we restart the xenlite service.

When we move to a later lite agent which uses IOCTLs rather than
WMI, we can remove this workaround