CA-258622: Fix unhandled exception in Helpers.ServiceRestart()
A 'ServiceController' object can be created with a non-existent
service. If this is the case, a call to the 'CanStop()' member
method of the object will throw an 'InvalidOperationException'.

Put the 'CanStop()' check in a try/catch block.

Also, add a call to 'Refresh()' before we check the 'Status'
field and replace the generic 'Exception' with 'Win32Exception',
since this is the only exception other than 'InvalidOperationException'
that 'Start()'/'Stop()' can throw.

Signed-off-by: Kostas Ladopoulos <konstantinos.ladopoulos@citrix.com>